Update
Project is going slower than I would like; however, a method was
developed for preparing the bytes of a given file for sending, and the
FileChooser is coming along nicely. Next steps will be: to finish/test
the FileChooser, integrate and test the FileChooser with the GUI, then
turn to the mobile app for working with its protocol, and expand the
scope of its file system.